Move icon service into services folder.
