+++ modules/custom/az_finder/az_finder.services.yml
@@ -2,7 +2,7 @@ services:
   _defaults:
     autoconfigure: true
   az_finder.icons:
-    class: Drupal\az_finder\AZFinderIcons
+    class: Drupal\az_finder\Service\AZFinderIcons
   az_finder.view_options:
     class: Drupal\az_finder\Service\AZFinderViewOptions
     arguments:
